requirements.txt
# Document Processing
docling>=2.0.0
docling-core>=2.0.0

# Vector Database
lancedb>=0.3.0
pyarrow>=14.0.0

# Embeddings
openai>=1.0.0
tiktoken>=0.5.0
transformers>=4.30.0

# MCP Framework
mcp>=0.9.0

# Citation Metadata
requests>=2.31.0
pybtex>=0.24.0
pdf2bib>=1.0.0

# PDF Processing
pymupdf>=1.23.0

# HTTP Client (for remote upload script)
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0

# Logging
rich>=13.0.0

# Type hints
typing-extensions>=4.0.0